cookbook/pocketflow-mcp/requirements.txt
pocketflow>=0.0.1
openai>=1.0.0
fastmcp # 用于 Model Context Protocol (MCP)
pyyaml # 用于处理 YAML 配置文件
